Move some compile task test cases
