+++ subprojects/language-c/src/test/groovy/dev/nokee/language/c/CCompileTaskIntegrationTest.java
@@ -31,6 +31,7 @@ import static dev.nokee.internal.testing.GradleNamedMatchers.named;
 import static dev.nokee.language.nativebase.internal.NativePlatformFactory.create;
 import static dev.nokee.runtime.nativebase.internal.TargetMachines.of;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
 @PluginRequirement.Require(id = "dev.nokee.c-language-base")
 class CCompileTaskIntegrationTest extends AbstractPluginTest implements CCompileTester {
@@ -56,5 +57,10 @@ class CCompileTaskIntegrationTest extends AbstractPluginTest implements CCompile
 	@Test
 	void hasDescription() {
 		assertThat(subject, TaskMatchers.description("Compiles the sources ':mufa'."));
+	}
+
+	@Test
+	void enablesPositionIndependentCodeByDefault() {
+		assertThat(subject.isPositionIndependentCode(), is(true));
 	}
 }
