+++ subprojects/language-cpp/src/test/groovy/dev/nokee/language/cpp/CppCompileTaskIntegrationTest.java
@@ -31,6 +31,7 @@ import static dev.nokee.internal.testing.GradleNamedMatchers.named;
 import static dev.nokee.language.nativebase.internal.NativePlatformFactory.create;
 import static dev.nokee.runtime.nativebase.internal.TargetMachines.of;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
 @PluginRequirement.Require(id = "dev.nokee.cpp-language-base")
 class CppCompileTaskIntegrationTest extends AbstractPluginTest implements CppCompileTester {
@@ -56,5 +57,10 @@ class CppCompileTaskIntegrationTest extends AbstractPluginTest implements CppCom
 	@Test
 	void hasDescription() {
 		assertThat(subject, TaskMatchers.description("Compiles the sources ':gali'."));
+	}
+
+	@Test
+	void enablesPositionIndependentCodeByDefault() {
+		assertThat(subject.isPositionIndependentCode(), is(true));
 	}
 }
